refactor(backend): decompose chat orchestrator pipeline stages into focused modules
